refactor rpc_module
